feat: add domain repository, and separate models
