+++ src/application/services/interpolator/interpolatorInterface.ts
@@ -1,4 +1,4 @@
-import { Coord } from '@/domain/datasetInterface'
+import { Coord } from '@/domain/models/dataset/datasetInterface'
 import { HTMLCanvas } from '@/presentation/dom/HTMLCanvas'
 
 export interface InterpolatorInterface {
