+++ src/application/strategies/extractStrategies/symbolExtractByArea.ts
@@ -1,6 +1,6 @@
 import ExtractStrategyInterface from './extractStrategyInterface'
 import { ExtractParent } from './extractParent'
-import { Coord } from '@/domain/datasetInterface'
+import { Coord } from '@/domain/models/dataset/datasetInterface'
 
 export default class SymbolExtractByArea
   extends ExtractParent
