+++ src/domain/models/dataset/dataset.ts
@@ -1,4 +1,4 @@
-import { Vector } from './axes/axesInterface'
+import { Vector } from '../../repositories/axes/axesInterface'
 import { DatasetInterface, Plots, Plot, Coord } from './datasetInterface'
 
 export class Dataset implements DatasetInterface {
