+++ src/presentation/components/Canvas/CanvasAxis.vue
@@ -114,3 +114,4 @@ export default defineComponent({
   methods: {},
 })
 </script>
+@/domain/models/axis/axisInterface
