+++ src/presentation/components/Canvas/CanvasPlot.vue
@@ -159,3 +159,4 @@ export default defineComponent({
   },
 })
 </script>
+@/domain/models/dataset/datasetInterface
