+++ src/presentation/components/Export/DataTable.vue
@@ -114,3 +114,4 @@ export default defineComponent({
   },
 })
 </script>
+@/domain/models/dataset/datasetInterface
