+++ src/presentation/components/Magnifier/MagnifierPlots.vue
@@ -144,3 +144,4 @@ export default defineComponent({
   },
 })
 </script>
+@/domain/models/dataset/datasetInterface
